Fixed Blueprint
Summary: Made many changes to the code that enables it in Blueprint without causing crashes.

Detail:

Made a lot of changes that may or may not correspond to the coding style preference. So just take it or leave it.

I made it so the model component can be added to an actor through the component add. Additionally, I made it flexible to work with any actor type than the model actor.

I also removed some of the components ability to be added through the component add menu. Mainly because I made them be created through the model component instead when you set the corresponding files. Setting the files will create the baseline for the model. You can then also add the other components through the drop down if needed. Made it so they should not crash even if the model component has not been added yet.